+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - adds version-based index mapping update support to the Search Relevance plugin [#344](https://github.com/opensearch-project/search-relevance/pull/344)
 - LLM Judgement Customized Prompt Template Implementation  [#264](https://github.com/opensearch-project/search-relevance/pull/264)
 - Add `_search` endpoint for searching for Search Configurations using OpenSearch DSL  [#372](https://github.com/opensearch-project/search-relevance/pull/372)
+- Add `_search` endpoint for searching for Judgments using OpenSearch DSL  [#371](https://github.com/opensearch-project/search-relevance/pull/371)
 
 ### Enhancements
 - Add BWC and Integration Tests for Index Mapping Update with Schema Version [#349](https://github.com/opensearch-project/search-relevance/pull/349)